refactor: move cross-drive fallback test
