+++ tests/unit/test_publishing_workflow.py
@@ -988,3 +988,62 @@ class TestAutoPriceReductionDispatch:
             await test_bot.publish_ad(str(tmp_path / "ad.yaml"), ad_cfg, ad_cfg_orig, [], AdUpdateStrategy.MODIFY)
             mock_apply.assert_called_once()
             assert mock_apply.call_args.kwargs["mode"] == AdUpdateStrategy.MODIFY
+
+    @pytest.mark.asyncio
+    async def test_cross_drive_path_fallback_windows(
+        self, test_bot:KleinanzeigenBot, base_ad_config:dict[str, Any]
+    ) -> None:
+        """Test that cross-drive path handling falls back to absolute path on Windows."""
+        # Create ad config
+        ad_cfg = Ad.model_validate(
+            base_ad_config
+            | {
+                "updated_on": "2024-01-01T00:00:00",
+                "created_on": "2024-01-01T00:00:00",
+                "auto_price_reduction": {"enabled": True, "strategy": "FIXED", "amount": 10, "min_price": 50, "delay_reposts": 0, "delay_days": 0},
+                "price": 100,
+                "repost_count": 1,
+                "price_reduction_count": 0,
+            }
+        )
+        ad_cfg.update_content_hash()
+        ad_cfg_orig = ad_cfg.model_dump()
+
+        # Simulate Windows cross-drive scenario
+        # Config on D:, ad file on C:
+        test_bot.config_file_path = "D:\\project\\config.yaml"
+        ad_file = "C:\\temp\\test_ad.yaml"
+
+        # Create a sentinel exception to abort publish_ad early
+        class _SentinelException(Exception):
+            pass
+
+        # Track what path argument __apply_auto_price_reduction receives
+        recorded_path:list[str] = []
+
+        def mock_apply_auto_price_reduction(
+            ad_cfg:Ad,
+            ad_cfg_orig:dict[str, Any],
+            ad_file_relative:str,
+            *,
+            mode:AdUpdateStrategy = AdUpdateStrategy.REPLACE,
+        ) -> None:
+            _ = mode
+            recorded_path.append(ad_file_relative)
+            raise _SentinelException("Abort early for test")
+
+        with (
+            patch("kleinanzeigen_bot.price_reduction.apply_auto_price_reduction", side_effect = mock_apply_auto_price_reduction),
+            patch.object(test_bot, "web_open", new_callable = AsyncMock),
+            patch("kleinanzeigen_bot.delete_flow.delete_ad", new_callable = AsyncMock),
+        ):
+            # Call publish_ad and expect sentinel exception
+            try:
+                await test_bot.publish_ad(ad_file, ad_cfg, ad_cfg_orig, [], AdUpdateStrategy.REPLACE)
+                pytest.fail("Expected _SentinelException to be raised")
+            except _SentinelException:
+                pass
+
+        # Verify the path argument is the absolute path (fallback behavior)
+        assert len(recorded_path) == 1
+        assert recorded_path[0] == ad_file, f"Expected absolute path fallback, got: {recorded_path[0]}"
